#16973 (fixed) Unable to lauch test in CI.

+++ webapp/config/application.rb
@@ -62,11 +62,13 @@ module Webapp
     config.exceptions_app = self.routes
 
     config.after_initialize do
-      # Remove running entities imports on app boot.
-      # In order to fix locked entities imports.
-      # Check database connection which do not exist on assets compilation rake task.
-      connected = ::ActiveRecord::Base.connection_pool.with_connection(&:active?) rescue false
-      EntitiesImport.running.delete_all if connected
+      unless Rails.env.test?
+        # Remove running entities imports on app boot.
+        # In order to fix locked entities imports.
+        # Check database connection which do not exist on assets compilation rake task.
+        connected = ::ActiveRecord::Base.connection_pool.with_connection(&:active?) rescue false
+        EntitiesImport.running.delete_all if connected
+      end
     end
   end
 end
